BREAKING: The type "decimal" has been renamed to "float" (including supporting binding options).

+++ src/ts/data/enum.ts
@@ -25,7 +25,7 @@ export enum DataType {
     null = "null",
     function = "function",
     boolean = "boolean",
-    decimal = "decimal",
+    float = "float",
     number = "number",
     bigint = "bigint",
     string = "string",
